Add files via upload
Terraforming Mars gamelogs are long, large json files that are hard to analyse in large amounts. I wanted to write a script that helps me (and other players) to create a single Excel file from all the gamelogs. The main goal of this output file is to extract all production data from all games for all players - so scoring efficiency can be easily determined - and compared to benchmark data